% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/betaSharedTest.R
\name{betaSharedTest}
\alias{betaSharedTest}
\title{Beta shared test}
\usage{
betaSharedTest(tree, gene.data, colSpecies = colnames(gene.data),
  cores = 1)
}
\arguments{
\item{tree}{Species phylogeny (phylo object)}

\item{gene.data}{A matrix of expression values with samples in columns and genes in rows}

\item{colSpecies}{A character vector with same length as columns in gene.data, specifying the
species for the corresponding column.}

\item{cores}{Number of parallel processes to run}
}
\value{
List with:
\itemize{
  \item indivBetaRes: results from \code{\link{fitOneTheta}}
  \item sharedBetaRes: results from \code{\link{fitSharedBeta}}
  \item sharedBeta: the estimated shared beta parameter
  \item LRT: log likelihood ratio test statistic between the individual and shared beta model
}
}
\description{
Beta shared test
}
