% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EVEcore.R
\name{calcCovMatOU}
\alias{calcCovMatOU}
\title{Covariance from phylogeny under the OU model}
\usage{
calcCovMatOU(tree, alphas, evol.variance)
}
\arguments{
\item{tree}{Species phylogeny (phylo object)}

\item{alphas}{Vector of alpha parameters for each edge in the tree}

\item{evol.variance}{}
}
\value{
Covariance matrix
}
\description{
Calculates the OU model covariance matrix given alpha parameters for each edge
and the evolutionary variance at each node of a tree.
}
