% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EVEcore.R
\name{calcExpVarOU}
\alias{calcExpVarOU}
\title{Expectation and variance from phylogeny under the OU model}
\usage{
calcExpVarOU(tree, thetas, alphas, sigma2s, rootVar, rootE)
}
\arguments{
\item{tree}{Species phylogeny (phylo object)}

\item{thetas}{Vector of theta parameters for each edge in the tree}

\item{alphas}{Vector of alpha parameters for each edge in the tree}

\item{sigma2s}{Vector of sigma2 parameters for each edge in the tree}

\item{rootVar}{The evolutionary variance at the root node}

\item{rootE}{The expected species mean at the root node}
}
\value{
data.frame with expected.mean and evol.variance for each node
}
\description{
Calculate expected species mean, and evolutionary variance for each node in the tree
given theta, sigma^2 and alpha parameters for each edge in the tree.
}
