% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EVEfast.R
\name{dmvnorm_nocheck}
\alias{dmvnorm_nocheck}
\title{Faster multivariate normal density calculation}
\usage{
dmvnorm_nocheck(x, sigma, mean)
}
\arguments{
\item{x}{vector or matrix of values}

\item{sigma}{covariance matrix}

\item{mean}{mean vector}
}
\value{
log P-value
}
\description{
This is the same as \code{\link[mvtnorm]{dmvnorm}} except with no check on input parameters to run faster.
}
