% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EVEcore.R
\name{EVEmodel}
\alias{EVEmodel}
\title{Generate multivariate normal distribution parameters for an EVE model}
\usage{
EVEmodel(tree, thetas, alphas, sigma2s, beta, rootVar, rootE, index.expand)
}
\arguments{
\item{tree}{Species phylogeny (phylo object)}

\item{thetas}{Vector of theta parameters for each edge in the tree}

\item{alphas}{Vector of alpha parameters for each edge in the tree}

\item{sigma2s}{Vector of sigma2 parameters for each edge in the tree}

\item{beta}{The beta parameter}

\item{rootVar}{The evolutionary variance at the root node}

\item{rootE}{The expected species mean at the root node}

\item{index.expand}{A vector mapping individual samples, i.e. columns in the 
gene expression matrix, to the species in the phylogeny}
}
\value{
List with covariance matrix (sigma) and expected species means (mean)
}
\description{
Generate multivariate normal distribution parameters for an EVE model
}
