% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EVEcore.R
\name{evolVarOU}
\alias{evolVarOU}
\title{Calculate evolutionary variance from phylogeny under the OU model}
\usage{
evolVarOU(tree, alphas, sigma2s, rootVar)
}
\arguments{
\item{tree}{Species phylogeny (phylo object)}

\item{alphas}{Vector of alpha parameters for each edge in the tree}

\item{sigma2s}{Vector of sigma2 parameters for each edge in the tree}

\item{rootVar}{The evolutionary variance at the root node}
}
\value{
Vector of variances for each node
}
\description{
Calculate expected species mean, and evolutionary variance for each node in the tree
given theta, sigma^2 and alpha parameters for each edge in the tree.
}
