% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EVEcore.R
\name{expectedMeanOU}
\alias{expectedMeanOU}
\title{Calculate expected species mean from phylogeny under the OU model}
\usage{
expectedMeanOU(tree, thetas, alphas, rootE)
}
\arguments{
\item{tree}{Species phylogeny (phylo object)}

\item{thetas}{Vector of theta parameters for each edge in the tree}

\item{alphas}{Vector of alpha parameters for each edge in the tree}

\item{rootE}{The expected species mean at the root node}
}
\value{
Vector of expected species means for each node in the tree
}
\description{
Calculate expected species mean, and evolutionary variance for each node in the tree
given theta, sigma^2 and alpha parameters for each edge in the tree.
}
