% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/betaSharedTest.R
\name{fitIndivBeta}
\alias{fitIndivBeta}
\title{Maximum likelihood estimation assuming an individual beta for each gene.}
\usage{
fitIndivBeta(tree, gene.data, colSpecies = colnames(gene.data))
}
\arguments{
\item{tree}{a phylogenic tree created from a .newick file}

\item{gene.data}{a matrix which contains gene data}

\item{colSpecies}{a vector which specifies the species for columns of the gene data matrix. By default this parameter
is set to the column names contained within the gene data matrix.}
}
\value{
Returns a list which contains a list for each gene of their parameter values, the calculated maximum
    likelihood, the number of optim iterations, and a message on the exit status of the optim routine
}
\description{
Calculates the parameters to achieve the maximum likelihood for a gene assuming that each gene has it's own individual
    beta value.
}
\details{
Calculates the maximum likelihood for a gene and its associated parameters through optimizing the original parameter
    estimates. This function assumes that each gene has its own individual beta value, and optimizes four parameters
    using optim, to find the parameters that give the greatest likelihood.
}
\examples{
test_res <- fitIndivBeta(tree, gene.data)
test_res <- fitIndivBeta(tree, gene.data, colSpecies = c("a", "b"))
}
