% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/betaSharedTest.R, R/oneThetaModel.R,
%   R/twoThetaModel.R
\name{fitSharedBeta}
\alias{fitSharedBeta}
\alias{fitOneTheta}
\alias{fitTwoTheta}
\title{Fit the maximum likelihood estimate of the parameters for a specific EVE model}
\usage{
fitSharedBeta(sharedBeta, tree, gene.data,
  colSpecies = colnames(gene.data), extra.var = NULL,
  lowerBound = c(theta = -99, sigma2 = 1e-04, alpha = 0.001),
  upperBound = c(theta = 99, sigma2 = 9999, alpha = 999),
  logTransPars = c("alpha", "sigma2", "beta"), cores = 1, fork = F)

fitOneTheta(tree, gene.data, colSpecies = colnames(gene.data),
  extra.var = NULL, lowerBound = c(theta = -99, sigma2 = 1e-04, alpha =
  0.001, beta = 0.001), upperBound = c(theta = 99, sigma2 = 9999, alpha =
  999, beta = 99), logTransPars = c("alpha", "sigma2", "beta"),
  cores = 1, fork = F)

fitTwoTheta(tree, gene.data, isTheta2edge,
  colSpecies = colnames(gene.data), extra.var = NULL,
  lowerBound = c(theta1 = -99, theta2 = -99, sigma2 = 1e-04, alpha =
  0.001, beta = 0.001), upperBound = c(theta1 = 99, theta2 = 99, sigma2 =
  9999, alpha = 999, beta = 99), logTransPars = c("alpha", "sigma2",
  "beta"), cores = 1, fork = F)
}
\arguments{
\item{sharedBeta}{the shared beta parameter}

\item{tree}{Phylogeny}

\item{gene.data}{A matrix of expression values with samples in columns and genes in rows}

\item{colSpecies}{A character vector with same length as columns in returned expression matrix, 
specifying the species, i.e. tip labels in the phylogeny, for the corresponding column.}

\item{extra.var}{An optional matrix of technical variance for each expression value. Same dimensions as gene.data (default=NULL)}

\item{lowerBound}{A named numeric vector of the lower bound for the estimated parameters}

\item{upperBound}{A named numeric vector of the upper bound for the estimated parameters}

\item{logTransPars}{A character vector with the names of the parameters that will be log transformed}

\item{cores}{Number of parallel processes to run}

\item{fork}{Use forking for parallel execution}

\item{isTheta2edge}{Logical vector with same length as number of edges in tree specifying whether
the corresponding edge theta parameter should be theta2 (TRUE) or theta1 (FALSE)}
}
\value{
A list with:
\describe{
  \item{par}{A matrix with the parameter estimates}
  \item{ll}{The log likelihood for each estimate}
  \item{iterations}{Number of iterations of the optimisation routine before reaching the estimate}
  \item{convergence}{Error code from the \code{\link[stats]{optim}} function. 0 means convergence.}
  \item{message}{Error message from the \code{\link[stats]{optim}} function.}
}
}
\description{
Fit the maximum likelihood estimate of the parameters for a specific EVE model
}
\section{Functions}{
\itemize{
\item \code{fitSharedBeta}: Fit model with a given shared beta for all genes

\item \code{fitTwoTheta}: Fit model with two different thetas assigned to specific edges
}}

