% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getEdgesFromMRCA.R
\name{getEdgesFromMRCA}
\alias{getEdgesFromMRCA}
\title{Get the edges for a clade in a phylogeny given the tips}
\usage{
getEdgesFromMRCA(tree, tips, includeEdgeToMRCA)
}
\arguments{
\item{tree}{Phylogeny}

\item{tips}{Name of the tips that defines the clade}

\item{includeEdgeToMRCA}{If TRUE, also include the edge leading to the root of the clade}
}
\value{
vector of edge indices
}
\description{
Get the edges for a clade in a phylogeny given the tips
}
