% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EVEcore.R
\name{initialParams}
\alias{initialParams}
\title{Estimate initial parameter values without using phylogeny}
\usage{
initialParams(gene.data, colSpecies)
}
\arguments{
\item{gene.data}{A matrix of expression values with samples in columns and genes in rows}

\item{colSpecies}{A character vector with same length as columns in gene.data, specifying the
species for the corresponding column.}
}
\value{
A matrix of initial parameters with the parameters in columns and genes in rows
}
\description{
Estimate initial parameter values without using phylogeny
}
