% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twoThetaModel.R
\name{initParamsTwoTheta}
\alias{initParamsTwoTheta}
\title{Estimate initial parameter values for two-theta model without using phylogeny}
\usage{
initParamsTwoTheta(gene.data, colSpecies, shiftSpecies)
}
\arguments{
\item{gene.data}{A matrix of expression values with samples in columns and genes in rows}

\item{colSpecies}{A character vector with same length as columns in gene.data, specifying the
species for the corresponding column.}

\item{shiftSpecies}{character vector with species that are assigned to theta2}
}
\value{
A matrix of initial parameters with the parameters in columns and genes in rows
}
\description{
Estimate initial parameter values for two-theta model without using phylogeny
}
