% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/betaSharedTest.R
\name{LLPerGeneIndivBeta}
\alias{LLPerGeneIndivBeta}
\title{Likelihood for a gene assuming an individual beta.}
\usage{
LLPerGeneIndivBeta(param.matrix.row, tree, gene.data.row, index.expand)
}
\arguments{
\item{param.matrix.row}{Numeric vector that contains the initial parameter estimates (theta, sigma2, alpha, beta)
for a gene}

\item{tree}{Phylogenic tree created from a .newick file}

\item{gene.data.row}{Numeric vector that contains expression data of a gene}

\item{index.expand}{Numeric which specifies the index on which to expand a covariance matrix}
}
\value{
A numeric value that represents the maximum likelihood of the gene.
}
\description{
This function calculates the log likelihood for an individual gene assuming the gene has an individual beta value.
}
\details{
Calculates the log likelihood for a gene with its own value for the beta parameter. The function calculates a log likelihood under
the OU model with four paramaters - theta, sigma squared, alpha, and beta. This function uses a model where each gene has
its own value for the beta parameter.
}
\examples{
optim(c(12.3, 6.1, .94, 3.6), fn = LLPerGeneIndivBeta,
tree, gene.data, index.expand)
optim(params, fn = LLPerGeneIndivBeta, tree, gene.data, index.expand)
}
