% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/betaSharedTest.R
\name{LLPerGeneSharedBeta}
\alias{LLPerGeneSharedBeta}
\title{Likelihood for a gene assuming a shared beta value.}
\usage{
LLPerGeneSharedBeta(param.matrix.row, betaShared, tree, gene.data.row,
  index.expand)
}
\description{
This function calculates the log likelihood for genes assuming that each
    gene has the same beta value; a shared beta value between all genes.
}
