% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EVEfast.R
\name{prepEVEmodel}
\alias{prepEVEmodel}
\title{Generate a EVEmodel function with pre-computed data for faster execution}
\usage{
prepEVEmodel(tree, index.expand, thetaIdx, alphaIdx, sigma2Idx, betaIdx)
}
\arguments{
\item{tree}{Species phylogeny (phylo object)}

\item{index.expand}{A vector mapping individual samples, i.e. columns in the 
gene expression matrix, to the species in the phylogeny}

\item{thetaIdx}{Vector of integers holding the index of the theta parameter for each edge}

\item{alphaIdx}{Vector of integers holding the index of the alpha parameter for each edge}

\item{sigma2Idx}{Vector of integers holding the index of the sigma2 parameter for each edge}

\item{betaIdx}{Vector of integers holding the index of the beta parameter}
}
\value{
a function with a single parameter holding a vector of parameters
}
\description{
Generate a EVEmodel function with pre-computed data for faster execution
}
