% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oneThetaModel.R
\name{simOneTheta}
\alias{simOneTheta}
\title{Simulate expression data for the one-theta EVE model, i.e. same parameters on all edges}
\usage{
simOneTheta(n, tree, colSpecies, theta, sigma2, alpha, beta)
}
\arguments{
\item{n}{Number of "genes" to simulate}

\item{tree}{Phylogeny}

\item{colSpecies}{A character vector with same length as columns in returned expression matrix, 
specifying the species, i.e. tip labels in the phylogeny, for the corresponding column.}

\item{theta}{Value of the theta parameter}

\item{sigma2}{Value of the sigma2 parameter}

\item{alpha}{Value of the alpha parameter}

\item{beta}{Value of the beta parameter}
}
\value{
Matrix of simulated gene expression values with samples in columns and genes in rows
}
\description{
Simulate expression data for the one-theta EVE model, i.e. same parameters on all edges
}
