% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twoThetaTest.R
\name{twoThetaTest}
\alias{twoThetaTest}
\title{Test for shift in theta at branch}
\usage{
twoThetaTest(tree, gene.data, isTheta2edge,
  colSpecies = colnames(gene.data), ...)
}
\arguments{
\item{tree}{Species phylogeny (phylo object)}

\item{gene.data}{A matrix of expression values with samples in columns and genes in rows}

\item{isTheta2edge}{Logical vector with same length as number of edges in tree specifying whether
the corresponding edge theta parameter should be theta2 (TRUE) or theta1 (FALSE)}

\item{colSpecies}{A character vector with same length as columns in gene.data, specifying the
species for the corresponding column.}

\item{...}{parameters passed to \code{\link{fitOneTheta}} and \code{\link{fitTwoTheta}}}
}
\value{
List with: indivBetaRes (results from \code{\link{fitOneTheta}})
\itemize{
  \item oneThetaRes: results from \code{\link{fitOneTheta}}
  \item twoThetaRes: results from \code{\link{fitTwoTheta}}
  \item LRT: log likelihood ratio test statistic between the two-theta and one-theta model
}
}
\description{
Test for shift in theta at branch
}
