% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/betaSharedTest.R
\name{fitIndivBeta}
\alias{fitIndivBeta}
\title{Maximum liklihood estimation of paramaters fit with individual beta values for each gene.}
\usage{
fitIndivBeta(tree, gene.data, colSpecies = colnames(gene.data))
}
\arguments{
\item{tree}{a phylogenic tree created from a .newick file}

\item{gene.data}{a matrix which contains gene data}

\item{colSpecies}{a vector which specifies the species for columns of the gene data matrix}
}
\description{
Maximum liklihood estimation of paramaters fit with individual beta values for each gene.
}
\examples{
test_res <- fitIndivBeta(tree, gene.data)
test_res <- fitIndivBeta(tree, gene.data, colSpecies = c("a", "b"))
}
